Search and trends tabs refactoring
