+++ youtubeanalyzer/trends.py
@@ -1,32 +1,19 @@
 import pycountry
 from PySide6.QtCore import (
     Qt,
-    QSortFilterProxyModel,
-    QItemSelection,
     QTimer
 )
 from PySide6.QtWidgets import (
     QApplication,
     QHBoxLayout,
-    QVBoxLayout,
-    QSplitter,
     QWidget,
     QLabel,
     QPushButton,
     QComboBox,
-    QTableView,
-    QTabWidget,
-    QMessageBox,
-    QSpinBox
-)
-from PySide6.QtCharts import (
-    QChart
+    QMessageBox
 )
 from youtubeanalyzer.defines import (
     app_name
-)
-from youtubeanalyzer.theme import (
-    Theme
 )
 from youtubeanalyzer.settings import (
     Settings
@@ -35,88 +22,23 @@ from youtubeanalyzer.engine import (
     YoutubeApiEngine
 )
 from youtubeanalyzer.model import (
-    ResultFields,
-    ResultTableModel
+    ResultFields
 )
 from youtubeanalyzer.widgets import (
     create_link_label,
     critial_detailed_message,
     TabWorkspaceFactory,
-    VideoDetailsWidget,
-    AnalyticsWidget
+    AbstractVideoTableWorkspace
 )
 
 
-class TrendsWorkspace(QWidget):
+class TrendsWorkspace(AbstractVideoTableWorkspace):
     def __init__(self, settings: Settings, parent: QWidget = None):
-        super().__init__(parent)
+        super().__init__(settings, parent)
+        self._search_limit_spin_box.setMaximum(200)
 
-        self._settings = settings
         self._selected_category = None
         self._loaded_category_id = None
-
-        h_layout = QHBoxLayout()
-        h_layout.addWidget(QLabel(self.tr("Category:")))
-        self._category_combo_box = QComboBox()
-        h_layout.addWidget(self._category_combo_box, 1)
-        h_layout.addSpacing(10)
-        h_layout.addWidget(QLabel(self.tr("Region:")))
-        self._region_combo_box = QComboBox()
-        for country in pycountry.countries:
-            self._region_combo_box.addItem(country.name, country.alpha_2)
-        h_layout.addWidget(self._region_combo_box)
-        self._region_combo_box.currentIndexChanged.connect(self._update_categories)
-
-        h_layout.addStretch(2)
-        self._request_limit_spin_box = QSpinBox()
-        self._request_limit_spin_box.setToolTip(self.tr("Set the result limit"))
-        self._request_limit_spin_box.setMinimumWidth(50)
-        self._request_limit_spin_box.setRange(2, 200)
-        self._request_limit_spin_box.setValue(10)
-        h_layout.addWidget(self._request_limit_spin_box)
-
-        h_layout.addStretch()
-        self._show_button = QPushButton(self.tr("Show"))
-        self._show_button.setToolTip(self.tr("Click to show trends"))
-        self._show_button.clicked.connect(self._on_show_trends_clicked)
-        h_layout.addWidget(self._show_button)
-
-        self.model = ResultTableModel(self)
-        self._sort_model = QSortFilterProxyModel(self)
-        self._sort_model.setSortRole(ResultTableModel.SortRole)
-        self._sort_model.setSourceModel(self.model)
-        self._table_view = QTableView(self)
-        self._table_view.setModel(self._sort_model)
-        self._table_view.setSelectionMode(QTableView.SelectionMode.SingleSelection)
-        self._table_view.setSelectionBehavior(QTableView.SelectionBehavior.SelectRows)
-        self._table_view.setSortingEnabled(True)
-        self._table_view.horizontalHeader().setSectionsMovable(True)
-        self._table_view.selectionModel().selectionChanged.connect(self._on_table_row_changed)
-
-        self._side_tab_widget = QTabWidget()
-
-        self._details_widget = VideoDetailsWidget(self.model, self)
-        self._side_tab_widget.addTab(self._details_widget, self.tr("Details"))
-
-        self._analytics_widget = AnalyticsWidget(self.model, self)
-        self._side_tab_widget.addTab(self._analytics_widget, self.tr("Analytics"))
-        self._analytics_widget.set_current_index_following(self._settings.get(Settings.AnalyticsFollowTableSelect))
-        if int(self._settings.get(Settings.Theme)) == Theme.Dark:
-            self._analytics_widget.set_charts_theme(QChart.ChartTheme.ChartThemeDark)
-        else:
-            self._analytics_widget.set_charts_theme(QChart.ChartTheme.ChartThemeLight)
-
-        self._main_splitter = QSplitter(Qt.Orientation.Horizontal)
-        self._main_splitter.addWidget(self._table_view)
-        self._main_splitter.addWidget(self._side_tab_widget)
-        self._main_splitter.setCollapsible(0, False)
-        self._main_splitter.setStretchFactor(0, 3)
-        self._main_splitter.setStretchFactor(1, 1)
-
-        v_layout = QVBoxLayout()
-        v_layout.addLayout(h_layout)
-        v_layout.addWidget(self._main_splitter)
-        self.setLayout(v_layout)
 
         QTimer.singleShot(0, self, self._update_categories)
 
@@ -143,44 +65,32 @@ class TrendsWorkspace(QWidget):
             self._loaded_category_id = None
 
     def load_state(self):
-        request_limit = int(self._settings.get(Settings.RequestLimit))
-        self._request_limit_spin_box.setValue(request_limit)
-        # Restore main splitter
-        splitter_state = self._settings.get(Settings.MainSplitterState)
-        if splitter_state and not splitter_state.isEmpty():
-            self._main_splitter.restoreState(splitter_state)
-        # Restore main table
-        table_header_state = self._settings.get(Settings.MainTableHeaderState)
-        if not table_header_state.isEmpty():
-            self._table_view.horizontalHeader().restoreState(table_header_state)
-        self._table_view.resizeColumnsToContents()
-        self._analytics_widget.set_current_chart_index(int(self._settings.get(Settings.LastActiveChartIndex)))
+        super().load_state()
         index = self._region_combo_box.findData(self._settings.get(Settings.TrendsRegion))
         if index >= 0:
             self._region_combo_box.setCurrentIndex(index)
         self._loaded_category_id = int(self._settings.get(Settings.TrendsVideoCategoryId))
 
     def save_state(self):
-        self._settings.set(Settings.RequestLimit, self._request_limit_spin_box.value())
-        self._settings.set(Settings.LastActiveChartIndex, self._analytics_widget.get_current_chart_index())
-        self._settings.set(Settings.MainTableHeaderState, self._table_view.horizontalHeader().saveState())
-        self._settings.set(Settings.MainSplitterState, self._main_splitter.saveState())
-        self._settings.set(Settings.LastActiveDetailsTab, self._side_tab_widget.currentIndex())
+        super().save_state()
         self._settings.set(Settings.TrendsRegion, self._region_combo_box.currentData())
         if self._category_combo_box.currentData():
             self._settings.set(Settings.TrendsVideoCategoryId, int(self._category_combo_box.currentData()))
 
-    def handle_preferences_change(self):
-        if int(self._settings.get(Settings.Theme)) == Theme.Dark:
-            self._analytics_widget.set_charts_theme(QChart.ChartTheme.ChartThemeDark)
-        else:
-            self._analytics_widget.set_charts_theme(QChart.ChartTheme.ChartThemeLight)
+    def _create_toolbar(self, h_layout: QHBoxLayout):
+        h_layout.addWidget(QLabel(self.tr("Category:")))
+        self._category_combo_box = QComboBox()
+        h_layout.addWidget(self._category_combo_box, 1)
+        h_layout.addSpacing(10)
+        h_layout.addWidget(QLabel(self.tr("Region:")))
+        self._region_combo_box = QComboBox()
+        for country in pycountry.countries:
+            self._region_combo_box.addItem(country.name, country.alpha_2)
+        h_layout.addWidget(self._region_combo_box)
+        self._region_combo_box.currentIndexChanged.connect(self._update_categories)
+        h_layout.addStretch(2)
 
-        self._analytics_widget.set_current_index_following(self._settings.get(Settings.AnalyticsFollowTableSelect))
-        if self._settings.get(Settings.AnalyticsFollowTableSelect):
-            self._analytics_widget.set_current_index(self._table_view.currentIndex())
-
-    def _on_show_trends_clicked(self):
+    def _on_search_clicked(self):
         self.setDisabled(True)
         QApplication.setOverrideCursor(Qt.CursorShape.BusyCursor)
 
@@ -209,7 +119,7 @@ class TrendsWorkspace(QWidget):
                                                          Please set it in the preferences"))
             return
 
-        request_limit = self._request_limit_spin_box.value()
+        request_limit = self._search_limit_spin_box.value()
         if not request_limit:
             request_limit = 10
             print("Request limit is not set. Using '10' by default")
@@ -240,16 +150,6 @@ class TrendsWorkspace(QWidget):
 
         QApplication.restoreOverrideCursor()
         self.setDisabled(False)
-
-    def _on_table_row_changed(self, current: QItemSelection, _previous: QItemSelection):
-        indexes = current.indexes()
-        if len(indexes) > 0:
-            index = self._sort_model.mapToSource(indexes[0])
-            self._details_widget.set_current_index(index)
-            self._analytics_widget.set_current_index(index)
-        else:
-            self._details_widget.set_current_index(None)
-            self._analytics_widget.set_current_index(None)
 
 
 class TrendsWorkspaceFactory(TabWorkspaceFactory):
